Added images and slightly edited the layout
All images are exported from the layout. Connected content images in the markup. Favicons connected to the project. Added width and height attributes to img tag in index.html and catalog.html. Removed the img tag after the username and added a link to the user's profile before the username in index.html and catalog.html