Code refactoring. Added text options, background colors and classes
Changed indentation from 4 spaces to two. Added text options for all elements and background colors. The states of interactive elements are described (text and background parameters). Custom fonts included. Added classes to markup. In index.html and catalog.html in the form, the img tag is replaced with svg for interactivity