Added aria-hidden="true" to svg tags. Button without text added aria-label, input field in the header of the site added label
Added the arial-hidden="true" attribute to SVG tags to prevent screen readers from seeing them. I added aria-label to the buttons in the slider in the promo block section in index.html. In index.html and catalog.html in the header added aria-label to the search button and a label for the search string